+++ CHANGELOG.md
@@ -13,6 +13,7 @@ This changelog is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.
 - Changed pytest version to "pytest>=8.3.4,<10" (#1917)
 
 ### Src
+- Updated `generated_proto.py` file to work with new proto version
 
 ### Examples
 
@@ -21,6 +22,7 @@ This changelog is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.
 ### Docs
 - Replaced relative documentation links in `README.md` with absolute GitHub URLs to fix broken PyPI rendering.
 - docs: Clarified AI usage in Good First Issues templates. (#1923)
+
 
 ### .github
 - archived workflows relating to PR reminders